[6] Now engine have a window class! But it's too far..
What am I talking about? There's a list of tons of features and classes and tech that should (and *MUST*) be added before very first engine's window will be created and displayed. So that, class re::desktop_window will not be used for a quite some time in order to create proper foundation beforehand and expand codebase to provide maybe a better desktop window solution. Code is tested and works well as expected, but won't be used for some time.